Got some results!

+++ Advanced Simulation - Assignment 1/Simulation/Simulation/Results/Result.java
@@ -3,13 +3,13 @@ package Simulation.Results;
 public class Result {
 
 	// General
-	private final int waitingTimeArbitraryCustomer;
-	private final int totalQueueLength;
+	private final double waitingTimeArbitraryCustomer;
+	private final double totalQueueLength;
 	private final double boatOccupancy;
 	
 	
-	public Result(	int waitingTimeArbitraryCustomer, 
-					int totalQueueLength,
+	public Result(	double waitingTimeArbitraryCustomer, 
+					double totalQueueLength,
 					double boatOccupancy)
 	{
 		this.waitingTimeArbitraryCustomer = waitingTimeArbitraryCustomer;
@@ -22,11 +22,11 @@ public class Result {
 		
 	}
 
-	public int GetWaitingTimeArbitraryCustomer() {
+	public double GetWaitingTimeArbitraryCustomer() {
 		return waitingTimeArbitraryCustomer;
 	}
 
-	public int GetTotalQueueLength() {
+	public double GetTotalQueueLength() {
 		return totalQueueLength;
 	}
